+++ static/js/actions.js
@@ -1,11 +1,11 @@
 import * as groupsActions from './ducks/groups';
 import * as userProfileActions from './ducks/userProfile';
-import * as sysinfoActions from './ducks/sysinfo';
+import * as sysInfoActions from './ducks/sysInfo';
 
 
 export default function hydrate() {
   return (dispatch) => {
-    dispatch(sysinfoActions.fetchSystemInfo());
+    dispatch(sysInfoActions.fetchSystemInfo());
     dispatch(userProfileActions.fetchUserProfile());
     dispatch(groupsActions.fetchGroups());
   };
