+++ static/js/components/NewTokenForm.jsx
@@ -2,7 +2,7 @@ import React, { useState } from 'react';
 import PropTypes from 'prop-types';
 import { useDispatch } from 'react-redux';
 
-import * as Action from '../ducks/groups';
+import * as Action from '../ducks/userProfile';
 
 
 const CreateTokenForm = ({ profile, groups }) => {
