+++ static/js/components/Plot.jsx
@@ -52,13 +52,12 @@ function bokeh_render_plot(node, docs_json, render_items, custom_model_js) {
 const Plot = (props) => {
   const dispatch = useDispatch();
   const plots = useSelector(state => state.plots);
-  const [state, setState] = useState({
-    error: false,
-    fetchingPlotIDs: []
-  });
+  const [error, setError] = useState(false);
+  const [fetchingPlotIDs, setFetchingPlotIDs] = useState([]);
+
   const needsFetching = () => (
     !plots.plotIDList.includes(props.url) &&
-    !state.fetchingPlotIDs.includes(props.url)
+    !fetchingPlotIDs.includes(props.url)
   );
   const fetchPlotDataIfNotCached = () => {
     if (needsFetching()) {
@@ -68,20 +67,13 @@ const Plot = (props) => {
           Actions.FETCH_SOURCE_PLOT
         )
       );
-      const { fetchingPlotIDs } = state;
-      setState({
-        ...state,
-        fetchingPlotIDs: fetchingPlotIDs.concat([props.url])
-      });
+      setFetchingPlotIDs(fetchingPlotIDs.concat([props.url]));
     }
     if (plots.plotData[props.url] &&
-        state.fetchingPlotIDs.includes(props.url)) {
-      const fetchingPlotIDs = state.fetchingPlotIDs.slice();
-      fetchingPlotIDs.splice(fetchingPlotIDs.indexOf(props.url), 1);
-      setState({
-        ...state,
-        fetchingPlotIDs
-      });
+        fetchingPlotIDs.includes(props.url)) {
+      const newFetchingPlotIDs = fetchingPlotIDs.slice();
+      newFetchingPlotIDs.splice(fetchingPlotIDs.indexOf(props.url), 1);
+      setFetchingPlotIDs(newFetchingPlotIDs);
     }
   };
   useEffect(() => {
@@ -90,7 +82,7 @@ const Plot = (props) => {
 
 
   const plotData = plots.plotData[props.url];
-  if (state.error) {
+  if (error) {
     return (
       <b>
         Error: Could not fetch plotting data
