+++ static/js/ducks/rotateLogo.js
