+++ static/js/ducks/sysInfo.js
